Player can shoot basic proyectiles, enemies dissappear when hp = 0

+++ Unity/MantaDive/Assets/Scripts/ScriptableObjectScripts/EnemyObjectScript.cs
@@ -7,6 +7,6 @@ public class EnemyObjectScript : ScriptableObject
     public Texture2D material;
     public bool followsPlayer;
     public float speed;
-    public float hitPoints;
+    public int hitPoints;
     public float damage;
 }
